+++ src/components/Header/AuthModal/Login.tsx
@@ -1,11 +1,12 @@
 import { X } from '@phosphor-icons/react';
+import { Link } from 'react-router-dom';
 
+import Modal from '@/components/Modal';
 import { AuthModalActions } from '@/types/auth';
 
-import Modal from '../Modal';
-
-export default function Register({ onClose, onToggle }: AuthModalActions) {
+export default function Login({ onClose, onToggle }: AuthModalActions) {
   const fieldStyle = 'flex flex-col gap-4 [&>label]:text-14';
+
   return (
     <>
       <Modal.Content>
@@ -17,17 +18,8 @@ export default function Register({ onClose, onToggle }: AuthModalActions) {
           >
             <X size={24} />
           </button>
-          <h2 className='text-24 font-semibold mb-24'>회원가입</h2>
-          <form className='flex flex-col gap-12'>
-            <div className={fieldStyle}>
-              <label htmlFor='name'>이름</label>
-              <input
-                id='name'
-                type='text'
-                className='input-primary'
-                placeholder='John Doe'
-              />
-            </div>
+          <h2 className='text-24 font-semibold mb-24'>로그인</h2>
+          <form className='flex flex-col gap-12 mb-4'>
             <div className={fieldStyle}>
               <label htmlFor='email'>이메일</label>
               <input
@@ -46,16 +38,10 @@ export default function Register({ onClose, onToggle }: AuthModalActions) {
                 placeholder='********'
               />
             </div>
-            <div className={fieldStyle}>
-              <label htmlFor='confrimPw'>비밀번호 확인</label>
-              <input
-                id='confrimPw'
-                type='password'
-                className='input-primary'
-                placeholder='********'
-              />
-            </div>
           </form>
+          <Link className='flex justify-end text-14 text-primary-300' to='/'>
+            비밀번호 찾기
+          </Link>
         </div>
       </Modal.Content>
       <Modal.Actions>
@@ -64,14 +50,14 @@ export default function Register({ onClose, onToggle }: AuthModalActions) {
           className='w-full h-45 px-24 bg-primary-300 text-15 
         text-white font-semibold rounded-md'
         >
-          가입하기
+          로그인
         </button>
         <button
           type='button'
           className='w-full text-14 text-primary-300'
           onClick={onToggle}
         >
-          이미 계정이 있으신가요? 로그인하기
+          계정 생성하기
         </button>
       </Modal.Actions>
     </>
